A hardware base solution for real-time leaf disease classification
Deep learning technologies hold transformative potential for the agricultural sector, particularly in enhancing food safety through advanced monitoring and management of crop health. In this paper, we explore the potential of edge computing for the real-time classification of leaf diseases using thermal imaging. We introduce a novel dataset of thermal images that encompasses a range of plant diseases. We also demonstrate the use of optimized deep learning models such as InceptionV3, MobileNetV1, MobileNetV2, and VGG-16, which employ pruning and quantization-aware training for deployment on resource-constrained devices like the Raspberry Pi 4b. Remarkably, these optimized models achieve inference times that surpass those of high-end GPUs such as the RTX3090, and they deliver state-of-the-art accuracy. These results demonstrate the potential of edge computing for time-sensitive applications.